#915 Add new case to command shell/run
I tried to delete a cache image but the character "*" of the command
was missing so the command fails.

Command received:
"rm -r /opt/opengnsys/cache/opt/opengnsys/images/test."
Command expected:
"rm -r /opt/opengnsys/cache/opt/opengnsys/images/test.*"

This command adds a new case to the call shell/run. This new case is
needed because in the case of deleting a image from the cache the web
only has to delete the last character of the string instead of last two
characters.