#942 Fix task break down
OpenGnsys tasks can contain other tasks. To know if a element of the
task is another task, we need to check if it has the parameter tareaid
(Task ID). The code had a bug and the check was done on an non-existent
parameter

This commit fix this bug changing the non-existent parameter for
tareaid.

Signed-off-by: OpenGnSys Support Team <soporte-og@soleta.eu>