Update command status
Now, without the ogAdmAgent, the WebConsole must updates the status of the
commands on the queue.

This commit changes the behaviour of the WebConsole to set the status of
the queued commands as finished correctly, if the communication with the
ogAdmServer was possible, and as finished with errors, if the
communication with the ogAdmServer was not possible.